refactor: 페이지의 복잡한 로직을 커스텀 훅 및 컴포넌트 분리
